feat: add zk verification to authorization state
This commit introduces a verifying key as optional attribute of the
authorization layer.

Every deployed program will be bound to an authorization contract, and
will optionally be a ZK program.

This commit introduces the skeleton for a ZK verification for such case.
It adds a proof binary to a Cosm execute message (the other variants are
currently unimplemented for ZK), and binds the authorization state to a
ZK key for the shielded program.

It uses ark-groth16, an industry standard that will support the output
proof of most of the zkVM implementations. As tech debt, we might want a
separated library that will decouple such decisions from the concrete
implementation, leaving all of the involved implementations (zkVM,
verifier, msg encoding) decoupled on this crate.